rename: BridgeGame 파일 model 폴더 안으로 이동
